+++ addons/website_event/tests/__init__.py
@@ -1,9 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import common
-from . import test_event_internals
-from . import test_ui
-from . import test_visitor_events
+from . import test_event_menus
+from . import test_event_visitor
 from . import test_website_event
-from . import test_website_event_access
