--- addons/website_event/tests/test_ui.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-import odoo.tests
-from odoo import tools
-
-
-@odoo.tests.tagged('post_install', '-at_install')
-class TestUi(odoo.tests.HttpCase):
-    def test_admin(self):
-        self.start_tour("/", 'website_event_tour', login='admin', step_delay=100)
