+++ addons/website_event_track/data/event_demo.xml
@@ -1,10 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <record id="event.event_0" model="event.event">
+        <field name="website_menu" eval="True"/>
         <field name="website_track" eval="True"/>
         <field name="website_track_proposal" eval="True"/>
     </record>
     <record id="event.event_7" model="event.event">
+        <field name="website_menu" eval="True"/>
         <field name="website_track" eval="True"/>
         <field name="website_track_proposal" eval="True"/>
     </record>
