+++ addons/website_event_track/models/event_event.py
@@ -74,7 +74,7 @@ class Event(models.Model):
         for event in self:
             if event.menu_id and (not menus_update_by_field or event in menus_update_by_field.get('website_track')):
                 event._update_website_menu_entry('website_track', 'track_menu_ids', 'track')
-            if event.menu_id and (not menus_update_by_field or event in menus_update_by_field.get('track')):
+            if event.menu_id and (not menus_update_by_field or event in menus_update_by_field.get('website_track_proposal')):
                 event._update_website_menu_entry('website_track_proposal', 'track_proposal_menu_ids', 'track_proposal')
 
     def _get_menu_type_field_matching(self):
