+++ addons/website_event_track/tests/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import test_event_menus
 from . import test_track_internals
-from . import test_website
 from . import test_website_event
